requirements.txt
coloredlogs==15.0.1
flatbuffers==25.9.23
matplotlib==3.10.8
huggingface_hub==1.2.3
humanfriendly==10.0
kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=f935c2ac0be6dd1bf96a5852973dad6103c48e866cf33532459fb1336bea4bfe
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.12.0.88
packaging==25.0
protobuf==6.33.2
pyctcdecode @ git+https://github.com/tz579/pyctcdecode@5964a111f0ea1d059b711c7cbae701d91e2e7b82
pyewts==0.2.0
pygtrie==2.5.0
pyreadline3==3.5.4
scipy==1.16.3
sympy==1.14.0
thin-plate-spline==1.2.0
boto3==1.42.10
asyncpg>=0.31.0
python-dotenv>=1.0.0
